# Fix/improve GetHugePageSize